Fix ScriptBase.py broken pipes causing application crash
Python pipes the stdout and stderr of subprocesses into the Python
shell stdout and stderr. When the Python script exits, the pipes close,
writes to stdout and stderr fail in the application, and the application
throws an exception and exits.

By using /dev/null instead, the application will never attempt writing
to a broken pipe.