Rename linux desktop files (#523)
The desktop file entry spec states that only alpha-numeric characters, dashes, and underscores are valid in desktop file names. The WPILib installer created files with spaces and parentheses, which caused issues for me when trying to launch any apps. This PR simply replaces spaces with underscores and removes the parentheses in just the file name, the actual display name is not changed.